Backend/TOPNET_PROJECT/participants/serializers.py
from rest_framework import serializers
from .models import Participant, Winner


class ParticipantSerializer(serializers.ModelSerializer):
    class Meta:
        model = Participant
        fields = ['id', 'first_name', 'last_name', 'email', 'phone', 'created_at']
        read_only_fields = ['created_at']


class WinnerSerializer(serializers.ModelSerializer):
    participant = ParticipantSerializer(read_only=True)

    class Meta:
        model = Winner
        fields = ['id', 'participant', 'draw_at']