Backend/TOPNET_PROJECT/requirements.txt
django==6.0.2
djangorestframework
django-cors-headers>=4.0
gunicorn>=22.0